net: change CNetAddr::ip to have flexible size #19628 (Part #3)

+++ src/netbase.cpp
@@ -12,6 +12,7 @@
 #include <util/system.h>
 
 #include <atomic>
+#include <limits>
 
 #ifndef WIN32
 #include <fcntl.h>
@@ -838,8 +839,8 @@ bool LookupSubNet(const std::string& strSubnet, CSubNet& ret)
         if (slash != strSubnet.npos)
         {
             std::string strNetmask = strSubnet.substr(slash + 1);
-            int32_t n;
-            if (ParseInt32(strNetmask, &n)) {
+            uint8_t n;
+            if (ParseUInt8(strNetmask, &n)) {
                 // If valid number, assume CIDR variable-length subnet masking
                 ret = CSubNet(network, n);
                 return ret.IsValid();
